remove duplicate stormPoint and debug console logs
